+++ mobile/lib/pages/lists/followers.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class FollowersScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Followers'),
+      ),
+    );
+  }
+}
