+++ mobile/lib/pages/lists/following.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class FollowingScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Following'),
+      ),
+    );
+  }
+}
